+++ diesel/src/macros/associations/mod.rs
@@ -1,2 +1,1 @@
 #[macro_use] mod belongs_to;
-#[macro_use] mod has_many;